changed rest of images to webp, moved some functions to other folders
