+++ about/index.html
@@ -103,10 +103,10 @@
             </div>
             <div class="image-wrapper">
               <img
-                src="../assets/img/boxed-water-is-better-d6V11M2Igak-unsplash.jpg"
+                src="../assets/img/boxed-water-is-better-d6V11M2Igak-unsplash.webp"
                 alt="person drinking from his bottle in a grassy hillside with mountains in the background" />
               <img
-                src="../assets/img/matthieu-petiard-Pf6e3o0GL4M-unsplash.jpg"
+                src="../assets/img/matthieu-petiard-Pf6e3o0GL4M-unsplash.webp"
                 alt="skier in the air with his skis crossed with blue skies behind him" />
             </div>
             <div class="btn-wrapper">
