+++ css/styles.css
@@ -644,10 +644,8 @@ footer {
 
 .card-btn:active,
 .card-btn:hover {
-  /* background-color: var(--dark-blue); */
   background-color: var(--dark-blue);
   color: var(--orange);
-
   transition: var(--transition);
 }
 
@@ -875,7 +873,7 @@ footer {
 
 /* ------- About me ------- */
 .about-top-img {
-  background: linear-gradient(rgba(0, 0, 0, 0), rgba(0, 0, 0, 0.1)), url(../assets/img/derek-sears-lzeMiXS3UJI-unsplash.jpg);
+  background: linear-gradient(rgba(0, 0, 0, 0), rgba(0, 0, 0, 0.1)), url(../assets/img/derek-sears-lzeMiXS3UJI-unsplash.webp);
   width: 100%;
   height: 400px;
   background-size: cover;
