+++ js/views/about.mjs
@@ -1,7 +1,7 @@
-import { toggleMobileNav } from "../ui/toggleMobileNav.mjs";
+import { toggleMobileNav } from "../helper/toggleMobileNav.mjs";
 toggleMobileNav();
-import { scrollCarouselByClick } from "../components/scrollCarouselByClick.mjs";
-import { scrollCarouselByDrag } from "../components/scrollCarouselByDrag.mjs";
+import { scrollCarouselByClick } from "../handler/scrollCarouselByClick.mjs";
+import { scrollCarouselByDrag } from "../handler/scrollCarouselByDrag.mjs";
 import { carousel, cardImages } from "../data/constants.mjs";
 import { cardHoverEffect } from "../ui/cardHoverEffect.mjs";
